fix_n: CardList 컴포넌트에서 Data fetch 함수  분리
