+++ src/utils/axios.js
@@ -1,0 +1,14 @@
+import axios from 'axios';
+
+export const request = async (method, url, data) => {
+  try {
+    const payload = await axios({
+      method,
+      url,
+      data,
+    });
+    return payload;
+  } catch (e) {
+    throw new Error(e);
+  }
+};
